+++ src/Symfony/Component/Debug/DebugClassLoader.php
@@ -259,6 +259,12 @@ class DebugClassLoader
 
         $parent = \get_parent_class($class);
         $parentAndOwnInterfaces = $this->getOwnInterfaces($class, $parent);
+
+        // at this moment, $parentAndOwnInterfaces contains only the class own interfaces.
+        if (isset($parentAndOwnInterfaces['Serializable']) && (!$refl->hasMethod('__serialize') || !$refl->hasMethod('__unserialize'))) {
+            $deprecations[] = sprintf('The "%s" %s the broken "\Serializable" interface. It is discouraged to do so. It is going to be deprecated and removed in future PHP versions.', $class, $refl->isInterface() ? 'interface extends' : 'class implements');
+        }
+
         if ($parent) {
             $parentAndOwnInterfaces[$parent] = $parent;